DSDEGP-1783 Fix IoUtil.getMd5HashOfGoogleObject when no md5 exists.
