+++ clio-client/src/main/scala/org/broadinstitute/clio/client/util/IoUtil.scala
@@ -5,10 +5,11 @@ import java.net.URI
 import java.nio.file.{FileVisitOption, Files, Path, Paths}
 import java.util.Comparator
 
-import scala.sys.process.Process
+class IoUtil private[util] (gsUtil: Option[GsUtil]) {
 
-trait IoUtil {
   val googleCloudStorageScheme = "gs"
+
+  val gs: GsUtil = gsUtil.getOrElse(new GsUtil(None))
 
   def isGoogleObject(location: URI): Boolean =
     Option(location.getScheme).contains(googleCloudStorageScheme)
@@ -47,35 +48,36 @@ trait IoUtil {
    * google-cloud-nio adapter will eagerly check storage authentication
    * for each path, and fail.
    */
-  import IoUtil.GsUtil
 
   def readGoogleObjectData(location: URI): String = {
-    val gs = new GsUtil(None)
     gs.cat(location.toString)
   }
 
   def copyGoogleObject(from: URI, to: URI): Int = {
-    val gs = new GsUtil(None)
     gs.cp(from.toString, to.toString)
   }
 
   def deleteGoogleObject(path: URI): Int = {
-    val gs = new GsUtil(None)
     gs.rm(path.toString)
   }
 
   def googleObjectExists(path: URI): Boolean = {
-    val gs = new GsUtil(None)
     gs.exists(path.toString) == 0
   }
 
-  def getMd5HashOfGoogleObject(path: URI): Symbol = {
-    Symbol(new GsUtil(None).md5Hash(path.toString))
+  private val md5HashPattern = "Hash \\(md5\\):\\s+([0-9a-f]+)".r
+
+  def getMd5HashOfGoogleObject(path: URI): Option[Symbol] = {
+    /*
+     * Files uploaded through parallel composite uploads won't have an md5 hash.
+     * See https://cloud.google.com/storage/docs/gsutil/commands/cp#parallel-composite-uploads
+     */
+    val rawHash = gs.hash(path.toString)
+    md5HashPattern.findFirstMatchIn(rawHash).map(m => Symbol(m.group(1)))
   }
 
   def getSizeOfGoogleObject(path: URI): Long = {
-    new GsUtil(None)
-      .du(path.toString)
+    gs.du(path.toString)
       .head
       .split("\\s+")
       .head
@@ -83,70 +85,8 @@ trait IoUtil {
   }
 
   def listGoogleObjects(path: URI): Seq[String] = {
-    new GsUtil(None).ls(path.toString)
+    gs.ls(path.toString)
   }
 }
-object IoUtil extends IoUtil {
-  // We should consider moving this to api usage instead of gsutil.
-  class GsUtil(stateDir: Option[Path]) {
 
-    def ls(path: String): Seq[String] = {
-      runGsUtilAndGetStdout(Seq("ls", path)).split("\n")
-    }
-
-    def du(path: String): Seq[String] = {
-      runGsUtilAndGetStdout(Seq("du", path)).split("\n")
-    }
-
-    def cp(from: String, to: String): Int = {
-      runGsUtilAndGetExitCode(Seq("cp", from, to))
-    }
-
-    def mv(from: String, to: String): Int = {
-      runGsUtilAndGetExitCode(Seq("mv", from, to))
-    }
-
-    def rm(path: String): Int = {
-      runGsUtilAndGetExitCode(Seq("rm", path))
-    }
-
-    def cat(objectLocation: String): String = {
-      runGsUtilAndGetStdout(Seq("cat", objectLocation))
-    }
-
-    def exists(path: String): Int = {
-      runGsUtilAndGetExitCode(Seq("-q", "stat", path))
-    }
-
-    def md5Hash(path: String): String = {
-      val output = runGsUtilAndGetStdout(Seq("hash", "-m", "-h", path))
-      output.lines.filter(_.contains("md5")).next.split(":")(1).trim
-    }
-
-    def runGsUtilAndGetStdout(gsUtilArgs: Seq[String]): String = {
-      val cmd = getGsUtilCmdWithStateDir(gsUtilArgs)
-      Process(cmd).!!
-    }
-
-    def runGsUtilAndGetExitCode(gsUtilArgs: Seq[String]): Int = {
-      val cmd = getGsUtilCmdWithStateDir(gsUtilArgs)
-      Process(cmd).!
-    }
-
-    def getGsUtilCmdWithStateDir(gsUtilArgs: Seq[String]): Seq[String] = {
-      def getTempStateDir: Path = {
-        val tempStateDir = Files.createTempDirectory("gsutil-save")
-        sys.addShutdownHook({
-          IoUtil.deleteDirectoryRecursively(tempStateDir)
-        })
-        tempStateDir
-      }
-
-      Seq(
-        "gsutil",
-        "-o",
-        "GSUtil:state_dir=" + stateDir.getOrElse(getTempStateDir)
-      ) ++ gsUtilArgs
-    }
-  }
-}
+object IoUtil extends IoUtil(None)
